refactor: svg에 색상을 적용할 수 있도록 설정 변경 및 적용
