+++ frontend/src/components/common/HeaderLogo/HeaderLogo.tsx
@@ -1,5 +1,5 @@
 import * as S from './HeaderLogo.style';
-import HeaderLogoImage from './HeaderLogo.svg';
+import HeaderLogoImage from '@/assets/HeaderLogo.svg';
 import ROUTES from '../../../constants/routes';
 
 function HeaderLogo() {
