+++ frontend/src/components/common/Rating/Rating.tsx
@@ -1,4 +1,5 @@
-import Heart from './heart.svg';
+import theme from '@/style/theme';
+import Heart from '@/assets/heart.svg';
 import * as S from './Rating.style';
 
 export type Props = {
@@ -11,7 +12,7 @@ function Rating({ rating, type = '일반', size = 'small' }: Props) {
   return type === '일반' ? (
     <S.Container>
       <S.Unit size={size}>
-        <Heart />
+        <Heart fill={theme.colors.primary} stroke={theme.colors.primaryDark} />
       </S.Unit>
       <S.Value size={size}>{rating.toFixed(2)}</S.Value>
     </S.Container>
@@ -19,7 +20,10 @@ function Rating({ rating, type = '일반', size = 'small' }: Props) {
     <S.Container>
       {Array.from({ length: rating }).map((_, index) => (
         <S.Unit key={index} size={size}>
-          <Heart />
+          <Heart
+            fill={theme.colors.primary}
+            stroke={theme.colors.primaryDark}
+          />
         </S.Unit>
       ))}
     </S.Container>
