+++ frontend/src/custom.d.ts
@@ -4,7 +4,7 @@ declare module '*.jpg' {
 }
 
 declare module '*.svg' {
-  const value: React.FC;
+  const value: React.FC<{ fill?: string; stroke?: string }>;
   export = value;
 }
